+++ .gitignore
@@ -3,7 +3,7 @@
 irods-basicGUI/logs/*
 irods-iBridgesGui.spec
 *DS_Store
-
+ibridgesgui/_version.py
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
