+++ ibridgesgui/gui_utils.py
@@ -1,9 +1,13 @@
 """Handy and reusable functions for the GUI"""
-import PyQt6
-import irods
 import pathlib
+from importlib.resources import files
 
+import irods
+import PyQt6
 from ibridges import get_collection, get_dataobject
+
+UI_FILE_DIR = files(__package__) / "ui_files"
+
 
 def populate_table(tableWidget, rows, data_by_row):
 
@@ -11,7 +15,7 @@ def populate_table(tableWidget, rows, data_by_row):
     for row, data in enumerate(data_by_row):
         for col, item in enumerate(data):
             tableWidget.setItem(row, col, PyQt6.QtWidgets.QTableWidgetItem(str(item)))
-    tableWidget.resizeColumnsToContents()    
+    tableWidget.resizeColumnsToContents()
 
 def get_irods_item(irods_path):
     try:
