+++ ibridgesgui/icons/refresh.png
