+++ ibridgesgui/ui_files/createCollection.ui
