+++ ibridgesgui/ui_files/irodsLogin.ui
