+++ ibridgesgui/ui_files/searchDialog.py
