+++ ibridgesgui/ui_files/searchDialog.ui
