+++ pyproject.toml
@@ -1,0 +1,79 @@
+[build-system]
+requires = ["setuptools>=45", "setuptools-scm[toml]>=6.2"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "ibridgesgui"
+authors = [
+    { name = "Christine Staiger", email = "c.staiger@uu.nl" },
+]
+description = "Graphical user interface for accessing data and metadata on iRods servers."
+readme = "README.md"
+requires-python = ">=3.8"
+keywords = ["irods", "data management", "storage"]
+license = { file = "LICENSE" }
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Development Status :: 3 - Alpha",
+    "License :: OSI Approved :: MIT License",
+]
+
+dependencies = [
+    "PyQt6>=6.4.2",
+    "ibridges",
+    # "pyinstaller==5.8.0",
+    "setproctitle==1.3.2",
+]
+
+dynamic = ["version"]
+
+[project.urls]
+GitHub = "https://github.com/chStaiger/iBridges-Gui"
+documentation = "https://github.com/chStaiger/iBridges-Gui"
+
+[project.optional-dependencies]
+test = [
+    "pylint",
+    "pytest",
+    "pytest-cov",
+    "ruff",
+    "mypy",
+]
+
+
+[project.scripts]
+ibridges-gui = "ibridgesgui.__main__:main"
+
+[tool.setuptools]
+packages = ["ibridgesgui"]
+
+[tool.setuptools_scm]
+write_to = "ibridgesgui/_version.py"
+
+[[tool.mypy.overrides]]
+module = [
+    "irods.*",
+]
+ignore_missing_imports = true
+
+[tool.pylint.'MASTER']
+ ignore-patterns="_version.py"
+
+[tool.pylint.'BASIC']
+good-names=["e", "f", "m"]
+
+
+[tool.pylint.'FORMAT']
+max-line-length=100
+max-locals=35
+max-args=10
+
+[tool.ruff]
+exclude = ["_version.py"]
+line-length = 100
+select = ["D", "E", "F", "W", "I", "N"]
+ignore = ["D203", "D213"]
